moved all non-state managing components into components folder
