+++ front-end/letsmeet/src/containers/Auth/SignUp/SignUp.js
@@ -70,11 +70,9 @@ const SignUp = (props) => {
         const nam = event.target.name;
         const val = event.target.value;
 
-        // 1 lowercase, 1 uppercase, 1 numeric, 1 special, and at least 8 chars
-
-
         if (nam === "password") {
             const isValid = val === authState.verifiedPassword.value;
+            // regex checks for: 1 lowercase, 1 uppercase, 1 numeric, 1 special, and at least 8 chars
             const pattern = new RegExp("^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#$%^&*])(?=.{8,})");
             const isValid2 = pattern.test(val);
 
@@ -154,11 +152,6 @@ const SignUp = (props) => {
                 // password strength not valid and passwords don't match
                 setErrorMessage(<strong>Your passwords do not match!</strong>);
             }
-
-            const pattern = new RegExp("^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#\$%\^&\*])(?=.{8,})");
-            const isValid2 = pattern.test(val);
-            (isValid2) ? setErrorMessage("") : setErrorMessage(<strong>Your passwords do not match or weak password!</strong>);
-
 
             const updatedVerifiedPassword = {
                 ...updatedAuthState[nam]
